% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineEdgeList.R
\name{combineEdgeList}
\alias{combineEdgeList}
\title{Combine edge list and database edge list.}
\usage{
combineEdgeList(con_mysql, new_edgelist, db_pmids,
  table_name = "edgelist", target_name = "target")
}
\arguments{
\item{con_mysql}{connection to mysql as defined in ~/.my.cnf}

\item{new_edgelist}{edge_list results, as obtained from `generateEdgeList` (see details)}

\item{db_pmids}{list of pmids used to search database.}

\item{table_name}{string with the name of the table used to store the edge list}

\item{target_name}{string with the name of the column used to store the target of edge list}
}
\value{
combineEdgeList() will return one dataframe after joining two datasets vertically.
}
\description{
`combineEdgeList` insert the edge list from the results of `generateEdgeList`
}
\details{
`combineEdgeList` must come from the result of generateEdgeList
}
\examples{

# place holder
}
