% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_date_table.R
\name{create_date_table}
\alias{create_date_table}
\title{Create edge list date table if not found.}
\usage{
create_date_table(conMysql, tableName = "EdgeList_date",
  targetName = "Target", dateName = "DatePMID")
}
\arguments{
\item{conMysql}{connection to mysql as defined in ~/.my.cnf}

\item{tableName}{string with the name of the table used to store the dates for pmids}

\item{targetName}{string with the name of the column used to store the Target of edge list}

\item{dateName}{string with the name of the column used to store the date of edge list}
}
\value{
create_date_table() always returns a scalar numeric that specifies the number of rows
        affected by the statement. An error is raised when issuing a statement over a closed
        or invalid connection, if the syntax of the statement is invalid, or if the statement is
        not a non-NA string.
}
\description{
`create_date_table` create edge list date table in MySQL
}
\details{
Default values for this table is: 
tableName = "EdgeList_date", targetName = "Target", dateName = "DatePMID"
MySQL statment:
CREATE TABLE Edgelist_date (
Target INT,
DatePMID date,
INDEX index_target(Target));
}
