% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_edge_list_table.R
\name{create_edge_list_table}
\alias{create_edge_list_table}
\title{Create edge list table if not found.}
\usage{
create_edge_list_table(con_mysql, tableName = "EdgeList",
  sourceName = "Source", targetName = "Target")
}
\arguments{
\item{con_mysql}{a MySQL connection}

\item{tableName}{string with the name of the table used to store the edge list, defaulting to 'EdgeList'}

\item{sourceName}{string with the name of the column used to store the source, defaulting to 'Source'}

\item{targetName}{string with the name of the column used to store the target, defaulting to 'Target'}
}
\value{
a scalar numeric that specifies the number of rows
        affected by the statement. An error is raised when issuing a statement over a closed
        or invalid connection, if the syntax of the statement is invalid, or if the statement is
        not a non-NA string.
}
\description{
`create_edge_list_table` create edge list table in MySQL.
}
\details{
Default values for this table is: 
tableName = "EdgeList", sourceName = "Source", targetName = "Target"
MySQL statment:
CREATE TABLE Edgelist_date (
SourcePMID INT,
Target INT,
INDEX index_target(Target));
}
\examples{

}
