% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_list_expansion.R
\name{edge_list_expansion}
\alias{edge_list_expansion}
\title{Collect additional citations from the supplied edge list.}
\usage{
edge_list_expansion(edge_list)
}
\arguments{
\item{edge_list}{edge_list results, as obtained from `generateEdgeList` (see details)}
}
\value{
An edge list (data.frame) with one column for target PMIDS and one column for source PMIDS.
}
\description{
`edge_list_expansion` insert the edge list from the results of `generateEdgeList`
}
\details{
`edge_list` must come from the result of generateEdgeList
}
\examples{

# Collect addition citations from the supplied 'e2'
res1 <- get_pmc_cited_in(21876761)
e2 <- generateEdgeList(res1)
edge_list_expansion(e2)
}
\seealso{
\code{\link{get_pmc_cited_in}} for obtaining elink citation results

\code{\link{generateEdgeList}} for obtaining elink citation results
}
