% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterEdgeList.R
\name{filterEdgeList}
\alias{filterEdgeList}
\title{Edge list filtering.}
\usage{
filterEdgeList(e, count)
}
\arguments{
\item{e}{an edge list, as obtained from `generateEdgeList`}
}
\value{
A list with the following objects:
        edgeList - the filtered edge list
        counts - a tibble with one column for target PMIDS and one column, 'n', containing the corresponding frequency of citations
}
\description{
`filterEdgeList` filters an edge list by citation threshold
}
\details{
This function returns an edge list having
target IDs with at least 'count' citations
}
\examples{

# generates an edge list for multiple articles
res <- get_pmc_cited_in(c(21876761, 311,29463753))
e <- generateEdgeList(res)
e2 <- filterEdgeList(e, 5)
}
\seealso{
\code{\link{generateEdgeList}} for generating an edge list, and 
         \code{\link{getCitationCounts}} for counting citation frequencies
}
