% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pmids_not_in_db.R
\name{find_pmids_not_in_db}
\alias{find_pmids_not_in_db}
\title{Find pmids that not in database}
\usage{
find_pmids_not_in_db(conMysql, pmids, tableName = "EdgeList_Date",
  targetName = "Target")
}
\arguments{
\item{conMysql}{connection to mysql as defined in ~/.my.cnf}

\item{pmids}{pmids a vector of PMIDs look-up}

\item{tableName}{string with the name of the table used to store the edge list}

\item{targetName}{string with the name of the column used to store the Target of edge list}
}
\value{
find_pmids_not_in_db() return the pmids that are not in the database.
}
\description{
`find_pmids_not_in_db` return the pmids that are not in the database.
}
\details{
`find_pmids_not_in_db` will search EdgeList_date for matching pmids and 
return the pmids that are not in the database.
}
\examples{
# In this example, PMIDs 123 and 456 are in the database
pmids_lookup <- c(123, 456, 759, 869)
res <- find_pmids_not_in_db(conMysql, pmids_lookup)
# res will print out 759 and 869
}
\seealso{
\code{\link{create_date_table}} for edge list date table structure
}
