% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEdgeList.R
\name{generateEdgeList}
\alias{generateEdgeList}
\title{Constructs edge lists from elink citation results.}
\usage{
generateEdgeList(elink)
}
\arguments{
\item{elink}{elink results, as obtained from `get_pmc_cited_in` (see details)}
}
\value{
An edge list (data.frame) with one column for target PMIDS and one column for source PMIDS.
}
\description{
`generateEdgeList` generates the edge list for results from `get_pmc_cited_in`
}
\details{
`elink` must be one of the following: a list containing one elink object, 
with list name corresponding to target PMID; a list of elink objects,
with names corresponding to target PMIDs, or a list of any combination
of the above. These corresond to a single PMID, a single batch,
and multiple batches, respectively.
}
\examples{

# generate an edge list for a single article
res1 <- get_pmc_cited_in(21876761)
e2 <- generateEdgeList(res1)

# generate an edge list for multiple articles
res2 <- get_pmc_cited_in(c(21876761, 311,29463753), batchSize = 2)
e2 <- generateEdgeList(res2)
}
\seealso{
\code{\link{get_pmc_cited_in}} for obtaining elink citation results
}
