% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_article_info.R
\name{get_article_info}
\alias{get_article_info}
\title{Get article information}
\usage{
get_article_info(pmids)
}
\arguments{
\item{pmids}{a vector of PMIDs look-up}
}
\value{
a data.frame containing article information for each PMID
}
\description{
`get_article_info` takes a vector of article PMIDs and returns summary 
information about the articles
}
\details{
Returns summary information for a list of articlees
This function is a wrapper for the `entrez_summary` and 
'extract_from_esummary' functions from `rentrez`. 
Function is currently limited to handling no more than 5000 PMIDs. 
The "web history" feature is used to store IDs on the 
server, and summary information is retrieved 500 IDs at a time.
Currently the 'uid', 'title', 'lastauthor', and 'fulljournalname'
for each PMID is obtained.
}
\examples{
res <- get_article_info(c(21876726,21876761))
}
