% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_list_db.R
\name{get_edge_list_db}
\alias{get_edge_list_db}
\title{Retrieve edge list from database}
\usage{
get_edge_list_db(con_mysql, pmids, table_name = "edgelist",
  target_name = "Target")
}
\arguments{
\item{con_mysql}{connection to mysql as defined in ~/.my.cnf}

\item{pmids}{list of pmids used to search database.}

\item{table_name}{string with the name of the table used to store the edge list}

\item{target_name}{string with the name of the column used to store the Target of edge list}
}
\value{
get_edge_list_db() will return the edge list in dataframe
}
\description{
`get_edge_list_db` will retrieve edge list from database with the provided list of PMIDs.
}
\details{
`get_edge_list_db` will retrieve edge list from database with the provided list of PMIDs.
}
\examples{
x <- c(21876761, 311, 29463753, 21876726)
get_edge_list_db(con_mysql, x)
}
