% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pmc_cited_in.R
\name{get_pmc_cited_in}
\alias{get_pmc_cited_in}
\title{Finds "cited in" articles from PMC.}
\usage{
get_pmc_cited_in(pmids, batchSize = 200)
}
\arguments{
\item{pmids}{a vector of PMIDs look-up}

\item{batchSize}{the batch size to use}
}
\value{
For looking up one PMID or one batch, an elink object; for multiple batches, a list of 
        of elink list objects are returned, i.e., the \eqn{i^{th}} element of the list is an elink list 
        object for the \eqn{i^{th}} batch.
}
\description{
`get_pmc_cited_in` takes a vector of article PMIDs and returns elink results containing
the PMIDs of articles that cite each of the given articles.
}
\details{
Finds articles that cite the provided set of articles. This
function is a wrapper for the `entrez_link` function from
`rentrez`. If more than 1 PMID is specified, requests are made in batches, ensuring that
no more than 200 pmids are queried at a time, and that no
more than 3 requests are made per second. If an Entrez Key
is set, this function allows for 10 requests per second.
Entrez keys can be set using `set_entrez_key(key)`.
}
\examples{
res <- get_pmc_cited_in(c(21876726,21876761))
}
