% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pmids_by_date.R
\name{get_pmids_by_date}
\alias{get_pmids_by_date}
\title{Query an pmids date from DB}
\usage{
get_pmids_by_date(conMysql, pmids, lastUpdate,
  tableName = "Edgelist_date")
}
\arguments{
\item{conMysql}{a MySQL connection}

\item{pmids}{a vector of pmids to look up}

\item{lastUpdate}{string of the date used for comparison, in YYYY-MM-DD format}

\item{tableName}{string with the name of the table that stores the target pmids and dates.}

\item{mysqlOperator}{string for comparison operators '= , >, >=, <, <=', defaulting to ">"}
}
\value{
a data.frame with one column for Target and one column for lastUpdated
}
\description{
`get_pmids_by_date` query edgelist_date table for PMIDs that match.
}
\details{
`get_pmids_by_date` queries the edgelist_date table for target PMIDs, that optionally have 
been updated since 'lastUpdate'. The user can specify the 'tableName', but this table must have
column names 'Target' and 'lastUpdated'. An error will be generated if 'tableName' does not exist.
}
\examples{

}
