% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pmids_date.R
\name{get_pmids_date}
\alias{get_pmids_date}
\title{Query an pmids date from DB}
\usage{
get_pmids_date(conMysql, lastUpdate, tableName = "edgelist_date",
  dateName = "DatePMID", mysqlOperator = "<")
}
\arguments{
\item{conMysql}{connection to mysql as defined in `~/.my.cnf`.}

\item{lastUpdate}{string of the date used for comparison.}

\item{tableName}{string with the name of the table used to store pmids and dates.}

\item{dateName}{string with the name of the column used to store the dates.}

\item{mysqlOperator}{string for comparison operators '= , >, >=, <, <='.}
}
\value{
An (data.frame) with one column for PMIDS and one column for date that pmid was saved to database.
}
\description{
`get_pmids_date` insert the edge list from the results of `generateEdgeList`
}
\details{
`get_pmids_date` must come from the result of generateEdgeList
}
\examples{


}
