% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCitationCounts.R
\name{getCitationCounts}
\alias{getCitationCounts}
\title{Counts the number of citations for each target ID.}
\usage{
getCitationCounts(e)
}
\arguments{
\item{e}{an edge list, as obtained from `generateEdgeList`}
}
\value{
A tibble with one column for target PMIDS and one column, 'n', containing the corresponding citation frequencies
}
\description{
`getCitationCounts` counts the number of citations for each target ID in an edge list
}
\details{
The edge list `e` must have a `Target` column. The number of citations for each
`Target` are counted. All citations per `Target` are assumed to be unique.
}
\examples{

# generates an edge list for multiple articles
res <- get_pmc_cited_in(c(21876761, 311,29463753))
e <- generateEdgeList(res)
counts <- getCitationCounts(e)
}
\seealso{
\code{\link{generateEdgeList}} for generating an edge list
}
