% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertEdgeList.R
\name{insertEdgeList}
\alias{insertEdgeList}
\title{Insert edge lists to mysql database.}
\usage{
insertEdgeList(con_mysql, edge_list, table_name = "EdgeList",
  source_name = "source", target_name = "Target",
  date_table_name = "EdgeList_date")
}
\arguments{
\item{con_mysql}{connection to mysql as defined in ~/.my.cnf}

\item{edge_list}{edge_list results, as obtained from `generateEdgeList` (see details)}

\item{table_name}{string with the name of the table used to store the edge list}

\item{source_name}{string with the name of the column used to store the source of edge list}

\item{target_name}{string with the name of the column used to store the Target of edge list}

\item{date_table_name}{string with the name of the table used to store the dates for pmids}
}
\value{
insertEdgeList() always returns a scalar numeric that specifies the number of rows
        affected by the statement. An error is raised when issuing a statement over a closed
        or invalid connection, if the syntax of the statement is invalid, or if the statement is
        not a non-NA string.
}
\description{
`insertEdgeList` insert the edge list from the results of `generateEdgeList`
}
\details{
`edge_list` must come from the result of generateEdgeList
}
\examples{

# generate an edge list for a single article and then insert the data to the table called "edgelist"
res1 <- get_pmc_cited_in(21876761)
e2 <- generateEdgeList(res1)
insertEdgeList(con_mysql, e2)
}
\seealso{
\code{\link{get_pmc_cited_in}} for obtaining edge_list citation results

\code{\link{generateEdgeList}} for obtaining edge_list citation results
}
