% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDate.R
\name{isDate}
\alias{isDate}
\title{isDate will check if date input work with MySQL}
\usage{
isDate(mydate)
}
\arguments{
\item{mydate}{string to be tested}

\item{date.format}{the date format}
}
\value{
isDate() returns true or false if the value is a date
}
\description{
`isDate` will try to format the input into a format recognizable to MySQL. If the formating fails, it will return NA
}
\details{
`edge_list` must come from the result of generateEdgeList
}
\examples{

# place holder


}
