% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLastUpdatedTable.R
\name{updateLastUpdatedTable}
\alias{updateLastUpdatedTable}
\title{Inserts Target and current date into MySQL database}
\usage{
updateLastUpdatedTable(conMysql, targetIDs, tableName = "EdgeList_date")
}
\arguments{
\item{conMysql}{a MySQL connection}

\item{targetIDs}{the target IDs to add to the table}

\item{tableName}{string with the name of the table used to store update dates, defaulting to 'EdgeList_date'#'}
}
\value{
a scalar numeric that specifies the number of rows
        affected by the statement. An error is raised when issuing a statement over a closed
        or invalid connection, if the syntax of the statement is invalid.
}
\description{
`updateLastUpdatedTable` inserts targetIDs and the current date into the specified MySQL table
}
\details{
The table will be created if it does not exist, with columns 'Target' and 'lastUpdated'
Targets form 'targetIDs' that are in the table will be removed.
}
\seealso{
\code{\link{generateEdgeList}} for obtaining edgeList citation results
}
