% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pmids_in_db.R
\name{check_pmids_in_db}
\alias{check_pmids_in_db}
\title{Check Edge List Database}
\usage{
check_pmids_in_db(con_mysql, pmids, table_name = "EdgeList",
  target_name = "Target_PMID")
}
\arguments{
\item{con_mysql}{connection to mysql as defined in ~/.my.cnf}

\item{table_name}{string with the name of the table used to store the edge list}

\item{target_name}{string with the name of the column used to store the target of edge list}

\item{edge_list}{edge_list results, as obtained from `generateEdgeList` (see details)}
}
\value{
check_pmids_in_db() return the pmids that are not already in the database.
}
\description{
`check_pmids_in_db` check if PMIDs have been searched and saved in database already. Only return PMIDs that are new.
}
\details{
`pmids` comes from user on shiny app
}
