% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineEdgeList.R
\name{combineEdgeList}
\alias{combineEdgeList}
\title{Combine 2 edge lists together}
\usage{
combineEdgeList(edgelist_1, edgelist_2)
}
\arguments{
\item{edgelist_1}{edge_list results, as obtained from `generateEdgeList` (see details)}

\item{edgelist_2}{edge_list results, as obtained from `generateEdgeList` (see details)}
}
\value{
combineEdgeList() will return one dataframe after joining two edge list vertically.
}
\description{
`edgelist_1` and `edgelist_2` must come from the result of `generateEdgeList` or `get_edge_list_db`
}
\details{
`edgelist_1` and `edgelist_2` must come from the result of `generateEdgeList` or `get_edge_list_db`.
If either edgelist_1 or edgelist_2 are NULL, only pass the edge list with values in it.
}
