% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_list_db.R
\name{get_edge_list_db}
\alias{get_edge_list_db}
\title{Retrieve edge list from database}
\usage{
get_edge_list_db(conMysql, pmids, tableName = "EdgeList")
}
\arguments{
\item{conMysql}{a MySQL connection}

\item{pmids}{list of target pmids used to search database.}

\item{tableName}{string with the name of the table used to store the edge list, defaulting to 'EdgeList'}
}
\value{
returns the edge list as a data.frame
}
\description{
retrieves an edge list from the database with the provided list of target PMIDs.
}
\details{
`get_edge_list_db` will retrieve an edge list from database with the provided list of target PMIDs.
An error is generated if the table does not exist in the database.
}
\examples{
# This will look for PMIDs in 'x' from the DB and get the Source and Target
x <- c(21876761, 311, 29463753, 21876726)
res <- get_edge_list_db(conMysql, x)
}
\seealso{
\code{\link{create_edge_list_table}} for edge list table structure
}
