% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_list_db.R
\name{get_edge_list_db}
\alias{get_edge_list_db}
\title{Retrieve edge list from database}
\usage{
get_edge_list_db(conMysql, pmids, tableName = "EdgeList",
  targetName = "Target")
}
\arguments{
\item{conMysql}{connection to mysql as defined in ~/.my.cnf}

\item{pmids}{list of pmids used to search database.}

\item{tableName}{string with the name of the table used to store the edge list}

\item{targetName}{string with the name of the column used to store the Target of edge list}
}
\value{
get_edge_list_db() will return the edge list in dataframe
}
\description{
`get_edge_list_db` will retrieve edge list from database with the provided list of PMIDs.
}
\details{
`get_edge_list_db` will retrieve edge list from database with the provided list of PMIDs.
If table is not found in database, NULL will be returned
}
\examples{
# This will look for PMIDs in 'x' from the DB and get the Source and Target
x <- c(21876761, 311, 29463753, 21876726)
res <- get_edge_list_db(conMysql, x)
}
\seealso{
\code{\link{create_edge_list_table}} for edge list table structure
}
