% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertEdgeList.R
\name{insertEdgeList}
\alias{insertEdgeList}
\title{Insert edge lists to mysql database.}
\usage{
insertEdgeList(conMysql, edgeList, tableName = "EdgeList",
  sourceName = "Source", targetName = "Target",
  dateTableName = "EdgeList_date")
}
\arguments{
\item{conMysql}{connection to mysql as defined in ~/.my.cnf}

\item{edgeList}{edgeList results, as obtained from `generateEdgeList` (see details)}

\item{tableName}{string with the name of the table used to store the edge list}

\item{sourceName}{string with the name of the column used to store the source of edge list}

\item{targetName}{string with the name of the column used to store the Target of edge list}

\item{dateTableName}{string with the name of the table used to store the dates for pmids}
}
\value{
insertEdgeList() always returns a scalar numeric that specifies the number of rows
        affected by the statement. An error is raised when issuing a statement over a closed
        or invalid connection, if the syntax of the statement is invalid, or if the statement is
        not a non-NA string.
}
\description{
`insertEdgeList` insert the edge list from the results of `generateEdgeList`
}
\details{
`edgeList` must come from the result of generateEdgeList. The edge list will be stored in to
table called Edgelist by default. This table have SourcePMID and Target column. It will also store
the date when the target PMID was inserted to the database in a table called EdgeList_date.
}
\examples{

# generate an edge list for a single article and then insert the data to the table called "edgelist"
res1 <- get_pmc_cited_in(21876761)
e2 <- generateEdgeList(res1)
insertEdgeList(conMysql, e2)
}
\seealso{
\code{\link{get_pmc_cited_in}} for obtaining edgeList citation results

\code{\link{generateEdgeList}} for obtaining edgeList citation results

\code{\link{create_edge_list_table}} for edge list table structure

\code{\link{create_date_table}} for edge list date table structure
}
