% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDate.R
\name{isDate}
\alias{isDate}
\title{isDate will format date input to work with MySQL}
\usage{
isDate(mydate)
}
\arguments{
\item{mydate}{string to be formated.}
}
\value{
isDate() returns the date in the format of YYYY-MM-DD.
}
\description{
`isDate` will try to format an input date into a format recognizable by MySQL, i.e., YYYY-MM-DD.
}
\details{
`isDate` will format the input as YYYY-MM-DD and will generate an error if the date cannot be formatted.
 The year should be specified as YYYY and must be >= 1900.
}
\examples{

# This will print 2001-01-09"
isDate("2001-1-9")

\dontrun{
# This will give an error
isDate("209-01-9999")
}
}
