% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDate.R
\name{isDate}
\alias{isDate}
\title{isDate will format date input to work with MySQL}
\usage{
isDate(mydate)
}
\arguments{
\item{mydate}{string to be tested}
}
\value{
isDate() returns true or false if the value is a date
}
\description{
`isDate` will try to format the input into a format recognizable to MySQL. If the formating fails, it will return NA
}
\details{
`isDate` must come from the result of generateEdgeList
}
\examples{

# This will print 2001-01-09"
date <- "2001-1-9"
<- isDate(date)

# This will return NA
date <- "209-01-9999"
res <- isDate(date)

}
