% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDate.R
\name{isDate}
\alias{isDate}
\title{isDate will format date input to work with MySQL}
\usage{
isDate(mydate)
}
\arguments{
\item{mydate}{string to be formated.}
}
\value{
isDate() returns date in the format of YYYY-MM-DD. Returns NA if it fails to format.
}
\description{
`isDate` will try to format the input into a format recognizable to MySQL. If the formating fails, it will return NA
}
\details{
`isDate` will try to format the input in YYYY-MM-DD and it will return NA if it cannot format the input.
}
\examples{

# This will print 2001-01-09"
date <- "2001-1-9"
<- isDate(date)

# This will return NA
date <- "209-01-9999"
res <- isDate(date)

}
