% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveEdgeList.R
\name{retrieveEdgeList}
\alias{retrieveEdgeList}
\title{Retrieve an edge list from the provided PMIDs}
\usage{
retrieveEdgeList(pmids, batchSize = 200, conMysql = NULL,
  lastUpdate = NULL)
}
\arguments{
\item{pmids}{a vector of PMIDs look-up.}

\item{batchSize}{the batch size to use.}

\item{conMysql}{connection to mysql as defined in `~/.my.cnf`.}

\item{lastUpdate}{string of a date used to find old PMIDs stored in database in YYYY-MM-DD format.}
}
\value{
An edge list (data.frame) with one column for target PMIDS and one column for source PMIDS.
}
\description{
`retrieveEdgeList` Retrieve an edge list from the provided PMIDs
}
\details{

}
\examples{

}
\seealso{
\code{\link{get_pmc_cited_in}} for obtaining edge_list citation results

\code{\link{generateEdgeList}} for obtaining edge_list citation results
}
